+++ site/configs/theme.ts
@@ -3,7 +3,7 @@ import type { Theme } from '@/types/configs/theme'
 
 const themeConfig: Theme = {
 	color: '#b96',
-	drawerWidth: 240
+	drawerWidth: 320
 }
 
 export default themeConfig