+++ site/types/components/molecules/profile.d.ts
@@ -1,0 +1,4 @@
+export type ProfileMoleculeProps = {
+	elevation?: number
+	borderRadius?: number
+}